+++ lib/src/curved_app_bar.dart
@@ -2,3 +2,4 @@ export 'clippers/curved_app_bar_clipper.dart';
 export 'enums/curved_app_bar_shape.dart';
 export 'models/curved_app_bar_action.dart';
 export 'widgets/curved_app_bar.dart';
+export 'widgets/curved_body.dart';